# Optimize ParseGatewayRDSRouteName